Added more prototype code

+++ ArcDataCore/Models/Sensor/SensorModel.cs
@@ -1,6 +1,6 @@
 ﻿namespace ArcDataCore.Models.Sensor
 {
-    public enum SensorModel : int
+    public enum SensorModel : ushort
     {
         As7262 = 100,
         Bme680 = 200,
